All Do/vinil_bot_kld/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-telegram-bot>=20.7
gspread>=5.12.0
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.108.0
aiocache>=0.12.2
aiohttp>=3.9.0
python-dotenv>=1.0.0
pillow>=10.1.0
imagehash>=4.3.1
pydantic>=2.5.0
httpx>=0.25.0
openai>=1.3.0
anthropic>=0.7.0
supabase>=2.0.0
pyjwt>=2.8.0
colorama>=0.4.6
